First pass at modularizing
